requirements.txt
absl-py==1.4.0
addict==2.4.0
aiohttp==3.8.5
aiosignal==1.3.1
albumentations==1.3.1
ansi2html==1.8.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
backcall==0.2.0
ballpark==1.4.0
beautifulsoup4==4.12.2
black==21.4b2
braceexpand==0.1.7
cachetools==5.3.0
ccimport==0.3.7
#certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.1.0
click==8.1.5
-e git+https://github.com/facebookresearch/ov-seg@36f49d496714998058d115ffb6172d9d84c59065#egg=clip&subdirectory=third_party/CLIP
cloudpickle==2.2.1
ConfigArgParse==1.5.5
cumm-cu111==0.2.9
cycler==0.11.0
Cython==3.0.0
dash==2.11.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.7
decorator==5.1.1
# detectron2==0.6+cu113
docker-pycreds==0.4.0
einops==0.6.1
entrypoints==0.4
fastjsonschema==2.17.1
filelock==3.12.2
fire==0.5.0
Flask==2.2.5
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
ftfy==6.1.1
future==0.18.3
fvcore==0.1.5.post20221221
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.18.1
google-auth-oauthlib==0.4.6
# Editable install with no version control (groundingdino==0.1.0)
# -e /home/tinhn/3D/ICCVW/GroundingDINO
grpcio==1.55.0
h5py==3.8.0
huggingface-hub==0.16.4
hydra-core==1.3.2
idna==3.4
imageio==2.28.1
imgviz==1.6.2
#importlib==1.0.4
#importlib-metadata==6.6.0
#importlib-resources==5.12.0
iopath==0.1.9
ipykernel==6.16.2
ipython==7.34.0
ipywidgets==8.0.7
# Editable install with no version control (isbnet==1.0)
# -e /home/tinhn/3D/ISBNet-mink
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-widgets==3.0.8
kiwisolver==1.4.4
kornia==0.6.12
lark==1.1.5
lightning-utilities==0.9.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
# MinkowskiEngine==0.5.4
# multidict==6.0.4
#MultiScaleDeformableAttention
munch==3.0.0
mypy-extensions==1.0.0
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.6.3
ninja==1.11.1
nltk==3.8.1
numpy==1.21.6
oauthlib==3.2.2
omegaconf==2.3.0
# -e git+https://github.com/KU-CVLAB/CAT-Seg@7b5e7a53a51c364feacbd68d0adf27ac651f6557#egg=open_clip_torch&subdirectory=open_clip
open3d==0.17.0
opencv-python==4.7.0.72
opencv-python-headless==4.8.0.74
packaging==23.1
pandas==1.3.5
parso==0.8.3
pathspec==0.11.1
pathtools==0.1.2
pccm==0.3.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
plotly==5.15.0
plyfile==0.9
# pointnet2 @ file:///home/tinhn/3D/ISBNet-mink/isbnet/pointnet2/dist/pointnet2-0.0.0-cp37-cp37m-linux_x86_64.whl
# pointops==1.0
portalocker==2.7.0
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.10.4
pycocotools==2.0.6
pydot==1.4.2
Pygments==2.15.1
pyparsing==3.1.0
pypng==0.20220715.0
pyquaternion==0.9.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-lightning==1.8.6
pytz==2023.3
pyviz3d==0.2.32
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==25.1.0
qudida==0.0.4
randomname==0.2.1
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rsa==4.9
safetensors==0.3.1
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
# Editable install with no version control (segment-anything==1.0)
# -e /home/tinhn/3D/ICCVW/segment_anything
sentencepiece==0.1.99
sentry-sdk==1.28.1
setproctitle==1.3.2
shapely==2.0.1
SharedArray==3.2.2
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
spconv-cu111==2.1.25
supervision==0.11.1
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.8.3.dev0
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
#torch==1.10.1+cu111
#torch-efficient-distloss==0.1.3
torch-ema==0.3
#torch-scatter==2.0.9
#torchaudio==0.10.1+cu111
#torchmetrics==0.11.4
#torchvision==0.11.2+cu111
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.2
transforms3d==0.4.1
trimesh==3.23.0
typed-ast==1.5.5
typing_extensions==4.7.1
urllib3==1.26.15
wandb==0.15.5
wcwidth==0.2.6
webdataset==0.2.48
Werkzeug==2.2.3
widgetsnbextension==4.0.8
yacs==0.1.8
yapf==0.40.1
yarl==1.9.2
zipp==3.15.0
